move virtual collection to the shared mobcat project
